arichitecture-poc/requirements.txt
fastapi
uvicorn
httpx
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp
grpcio